+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This release adds the following new features:
 - Added the ability to delete files in the file view by right clicking and selecting "Delete File".
 - Added a Kotlin lexer/parser for syntax highlighting.
 - Remember the last directory used for file operations and for opening conversations.
+- Renamed "workspace" to "mindspace".
 
 Te release fixes the following bugs: